requirements.txt
Flask==1.1.4
Flask-Sockets==0.2.1
Flask-Breadcrumbs==0.5.1
gevent==22.10.2
gevent-websocket==0.10.1
python-dotenv==0.21.0
requests==2.28.1
websocket-client==0.57.0
markupsafe==2.0.1
Pyrebase4==4.6.0